% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{annotReg}
\alias{annotReg}
\alias{annotReg,srnadiffExp-method}
\alias{annotReg<-}
\alias{annotReg<-,srnadiffExp-method}
\title{Accessors for the 'annotReg' slot of an srnadiffExp object}
\usage{
\S4method{annotReg}{srnadiffExp}(object)

\S4method{annotReg}{srnadiffExp}(object) <- value
}
\arguments{
\item{object}{An \code{srnadiffExp} object.}

\item{value}{Annotated regions as a \code{GRanges} object.}
}
\description{
The \code{annotReg} slot holds the annotated regions as a \code{GRanges}
object.
}
\examples{
srnaExp <- srnadiffExample()
annotReg(srnaExp)

}
