% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{bamFiles}
\alias{bamFiles}
\alias{bamFiles,srnadiffExp-method}
\title{Accessors for the 'bamFiles' slot of an srnadiffExp object}
\usage{
\S4method{bamFiles}{srnadiffExp}(object)
}
\arguments{
\item{object}{An \code{srnadiffExp} object.}
}
\description{
The \code{bamFiles} slot holds the full paths to the BAM files as a
\code{\link[Rsamtools]{BamFileList}}.
}
\examples{
require(Rsamtools)

srnaExp <- srnadiffExample()
path(bamFiles(srnaExp))

}
