% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{buildDataHmm}
\alias{buildDataHmm}
\title{Read the coverage and extract expressed regions}
\usage{
buildDataHmm(object, lengths, values)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{lengths}{The lengths of the RLE of the coverages: a \code{list} of \code{vector}s or \code{integer}s.}

\item{values}{The values of the RLE of the coverages: \code{list} of \code{vector}s or \code{integer}s.}
}
\value{
The selected values: a \code{list} of \code{vector}s of \code{integer}s.
}
\description{
Read the coverage and extract expressed regions
}
