% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{calcFactorRLE}
\alias{calcFactorRLE}
\title{Compute normalization factors using RLE values.}
\usage{
calcFactorRLE(object, counts, libSize)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{counts}{the merged coverage}

\item{libSize}{the library sizes}
}
\value{
The normalization factors
}
\description{
Compute normalization factors using RLE values.
}
