% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{chromosomeSizes}
\alias{chromosomeSizes}
\alias{chromosomeSizes,srnadiffExp-method}
\title{Accessors for the 'chromosomeSizes' slot of an srnadiffExp object}
\usage{
\S4method{chromosomeSizes}{srnadiffExp}(object)
}
\arguments{
\item{object}{An \code{srnadiffExp} object.}
}
\description{
The \code{chromosomeSizes} slot holds the sizes of the chromosomes
as a named vector with chromosome names.
}
\examples{
srnaExp <- srnadiffExample()
chromosomeSizes(srnaExp)

}
