% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeLogFoldChange}
\alias{computeLogFoldChange}
\title{Compute fold-change}
\usage{
computeLogFoldChange(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
The same object, with the fold change
}
\description{
Compute fold-change
}
