% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{computePvalues}
\alias{computePvalues}
\title{Compute p-values of the selected counts.}
\usage{
computePvalues(object, counts)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{counts}{The counts: a \code{list} of \code{vector}s or \code{integer}s.}
}
\value{
The p-values: a \code{list} of \code{numeric}.
}
\description{
Compute p-values of the selected counts.
}
