% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{countFeatures}
\alias{countFeatures}
\title{Count the number of reads in each condition in the given annotation.}
\usage{
countFeatures(object, annotation)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{annotation}{A \code{GRanges} object.}
}
\value{
A \code{list} of \code{vector} of \code{integer}s.
}
\description{
Count the number of reads in each condition in the given annotation.
}
