% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{coverages}
\alias{coverages}
\alias{coverages,srnadiffExp-method}
\title{Accessors for the 'coverages' slot of an srnadiffExp object}
\usage{
\S4method{coverages}{srnadiffExp}(object)
}
\arguments{
\item{object}{An \code{srnadiffExp} object.}
}
\description{
The \code{coverages} slot holds the coverages as a named
\code{\link[IRanges]{RleList}} object with one coverage vector
per seqlevel in the set of genomic alignments from the BAM files.
}
\examples{
srnaExp <- srnadiffExample()
coverages(srnaExp)

}
