% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{CPM}
\alias{CPM}
\title{Normalize counts}
\usage{
CPM(object, mergedCounts, libSize, factors)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{mergedCounts}{the counts: one per condition}

\item{libSize}{the library size}

\item{factors}{the normalization factors}
}
\value{
The normalized counts
}
\description{
Normalize counts
}
