% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{parameters}
\alias{parameters}
\alias{parameters,srnadiffExp-method}
\alias{parameters<-}
\alias{parameters<-,srnadiffExp-method}
\alias{print.srnadiff_par}
\title{Accessors for the 'parameters' slot of an srnadiffExp object}
\usage{
\S4method{parameters}{srnadiffExp}(object)

\S4method{parameters}{srnadiffExp}(object) <- value

\method{print}{srnadiff_par}(x, ...)
}
\arguments{
\item{object}{An \code{srnadiffExp} object.}

\item{value}{A named \code{list} containing valid parameters. See details.}

\item{x}{The first element of the parameters used by an \code{srnadiff}
object}

\item{...}{The other elements of the parameters}
}
\value{
The named list of the parameters used in the analysis.
}
\description{
The \code{parameters} slot holds the parameter values
used in an sRNA-diff approach as a named \code{list}. Default values
exist for parameters, but these can also be supplied as input
values in the \code{useParameters} argument of the \code{\link{srnadiff}}
function or using the assignment function \code{\link{parameters<-}}.
}
\details{

}
\examples{
srnaExp <- srnadiffExample()
srnaExp <- srnadiff(srnaExp)
print(parameters(srnaExp))

parameters(srnaExp) <- list("minSize" = 1, "maxSize" = 1500)

srnaExp <- srnadiffExample()
srnaExp <- srnadiff(srnaExp)
print(parameters(srnaExp))

}
\seealso{
\code{useParameters} argument in \code{\link{srnadiff}} function.
}
