% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRegion}
\alias{plotRegion}
\title{Plot a region}
\usage{
plotRegion(object, region)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{region}{A \code{GenomicRange} object.}
}
\value{
A \code{ggbio} object.
}
\description{
Plot a region
}
\examples{
dir         <- system.file("extdata", package="srnadiff", mustWork = TRUE)
data        <- read.csv(file.path(dir, "data.csv"))
gtfFile     <- file.path(dir, "Homo_sapiens.GRCh38.76.gtf")
annotation  <- readWholeGenomeAnnotation(gtfFile)
bamFiles    <- file.path(dir, data$FileName)
replicates  <- data$SampleName
conditions  <- factor(data$Condition)
exp         <- sRNADiffExp(annotation, bamFiles, replicates, conditions)
diffRegions <- runAll(exp)
plotRegion(exp, regions(diffRegions)[1])

}
