% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_buildHmm}
\alias{rcpp_buildHmm}
\title{Compute unique counts.}
\usage{
rcpp_buildHmm(coverages, minDepth)
}
\arguments{
\item{minDepth}{the minimum read coverage}

\item{lengths}{the sizes of the RLEs (one list per chromosome)}

\item{values}{the values of the RLEs (one list per chromosome)}
}
\value{
the unique counts
}
\description{
Compute unique counts.
}
