% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_clustering}
\alias{rcpp_clustering}
\title{Compute unique counts.}
\usage{
rcpp_clustering(lengths, values, chromosomeSizes, minDepth, minSize, maxSize,
  minDifference)
}
\arguments{
\item{lengths}{the sizes of the RLEs (one list per chromosome)}

\item{values}{the values of the RLEs (one list per chromosome)}

\item{chromosomeSizes}{the sizes of the chromosomes}

\item{minDepth}{minimum coverage}

\item{minSize}{minimum region size}

\item{maxSize}{maximum region size}

\item{minDifference}{minimum difference between 2 regions}
}
\value{
selected regions
}
\description{
Compute unique counts.
}
