% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_naive}
\alias{rcpp_naive}
\title{Compute naive method.}
\usage{
rcpp_naive(lengths, values, chromosomeSizes, depth, distance, size)
}
\arguments{
\item{lengths}{the sizes of the RLEs (one list per chromosome)}

\item{values}{the values of the RLEs (one list per chromosome)}

\item{chromosomeSizes}{the sizes of the chromosomes}

\item{depth}{minimum number of reads per position}

\item{distance}{threshold to merge consecutive regions}

\item{size}{minimum region size}
}
\value{
the unique counts
}
\description{
Compute naive method.
}
