% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_normalization}
\alias{rcpp_normalization}
\title{Normalize counts}
\usage{
rcpp_normalization(coverages, librarySizes)
}
\arguments{
\item{coverages}{the coverages (a list of RLEs)}

\item{librarySizes}{number of elements per sample}
}
\value{
the normalization factors
}
\description{
Normalize counts
}
