% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_viterbi}
\alias{rcpp_viterbi}
\title{Run the Viterbi algorithm on the HMM.}
\usage{
rcpp_viterbi(coverages, transitions, emissions, emissionThreshold, starts,
  counts, pvalues, minDepth, minSize, maxSize)
}
\arguments{
\item{transitions}{the transition log-probabilities}

\item{emissions}{the emission log-probabilities}

\item{emissionThreshold}{the emission threshold}

\item{starts}{the start log-probabilities}

\item{counts}{the unique counts}

\item{pvalues}{the p-values of the counts}

\item{minDepth}{the minimum read coverage}

\item{minSize}{the minimum size region}

\item{maxSize}{the maximum size region}

\item{lengths}{the sizes of the RLEs (one list per chromosome)}

\item{values}{the values of the RLEs (one list per chromosome)}
}
\value{
a segmentation of the chromosomes
}
\description{
Run the Viterbi algorithm on the HMM.
}
