% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{readMiRBaseMatureAnnotation}
\alias{readMiRBaseMatureAnnotation}
\title{Extraction of mature miRNAs using a miRBase-formatted file}
\usage{
readMiRBaseMatureAnnotation(fileName)
}
\arguments{
\item{fileName}{The annotation file name in GFF/GTF format.}
}
\value{
A GRanges.
}
\description{
The file should have the usual GFF3 miRBase format.
}
\examples{
dir        <- system.file("extdata", package="srnadiff", mustWork = TRUE)
gffFile    <- file.path(dir, "mirbase21_GRCh38.gff3")
annotation <- readMiRBaseMatureAnnotation(gffFile)

}
