% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{readWholeGenomeAnnotation}
\alias{readWholeGenomeAnnotation}
\title{Extraction of miRNAs using an GTF/GFF annotation file}
\usage{
readWholeGenomeAnnotation(fileName)
}
\arguments{
\item{fileName}{The annotation file name in GFF/GTF format.}
}
\value{
A GRanges.
}
\description{
Extract miRNAs regions from a GTF/GFF file.
The second field of the miRNAs should be \code{miRNA}.
}
\examples{
dir        <- system.file("extdata", package="srnadiff", mustWork = TRUE)
gtfFile    <- file.path(dir, "Homo_sapiens.GRCh38.76.gtf.gz")
annotation <- readWholeGenomeAnnotation(gtfFile)

}
