% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{reconcileRegions}
\alias{reconcileRegions}
\title{Keep regions with best p-values.}
\usage{
reconcileRegions(object, allSets)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{allSets}{A \code{GRanges} object.}
}
\value{
A \code{GRanges} object.
}
\description{
Keep regions with best p-values.
}
