% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{regions}
\alias{regions}
\alias{regions,sRNADiff-method}
\title{Get the output regions}
\usage{
regions(object)

\S4method{regions}{sRNADiff}(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
The selected regions
}
\description{
Get the output regions
}
\examples{
dir         <- system.file("extdata", package="srnadiff", mustWork = TRUE)
data        <- read.csv(file.path(dir, "data.csv"))
gtfFile     <- file.path(dir, "Homo_sapiens.GRCh38.76.gtf")
annotation  <- readWholeGenomeAnnotation(gtfFile)
bamFiles    <- file.path(dir, data$FileName)
replicates  <- data$SampleName
conditions  <- factor(data$Condition)
exp         <- sRNADiffExp(annotation, bamFiles, replicates, conditions)
regions(exp)

}
