% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{removeRedundant}
\alias{removeRedundant}
\title{Remove redundant regions}
\usage{
removeRedundant(regions, padj)
}
\arguments{
\item{regions}{An \code{GRanges} object.}

\item{padj}{A list of adjusted p-values.}
}
\value{
A \code{GRanges} object.
}
\description{
Remove redundant regions
}
