% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runAll}
\alias{runAll}
\title{Run the segmentation using different methods, and reconcile them.}
\usage{
runAll(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A \code{GRanges} object.
}
\description{
This is the main function of the code, and calls all the methods:
\itemize{
  \item find normalization factors;
  \item compute the log fold-change;
  \item call all the segmentation methods (annotation, HMM, IR);
  \item reconcile the regions found by the different methods.
}
}
\examples{
exp        <- sRNADiffExample()
exp        <- runAll(exp)

}
