% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{runAllAnnotation}
\alias{runAllAnnotation}
\title{Segmentation using an annotation file.}
\usage{
runAllAnnotation(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A GRanges.
}
\description{
Segmentation using an annotation file.
}
