% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runAllClustering}
\alias{runAllClustering}
\title{Segmentation of the genome using a clustering method.}
\usage{
runAllClustering(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A \code{GRanges} object.
}
\description{
Segmentation of the genome using a clustering method.
}
