% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{runAllHmm}
\alias{runAllHmm}
\title{Segmentation of the genome using an HMM.}
\usage{
runAllHmm(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A GRanges.
}
\description{
Segmentation of the genome using an HMM.
}
