% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive.R
\name{runAllNaive}
\alias{runAllNaive}
\title{Segmentation of the genome in a naive way.}
\usage{
runAllNaive(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A GRanges.
}
\description{
Segmentation of the genome in a naive way.
}
