% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{runAllSlice}
\alias{runAllSlice}
\title{Segmentation of the genome using a slice method.}
\usage{
runAllSlice(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A \code{GRanges} object.
}
\description{
Segmentation of the genome using a slice method.
}
