% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runClustering}
\alias{runClustering}
\title{Initialize and run the clustering.}
\usage{
runClustering(object, counts)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{counts}{The normalized counts}
}
\value{
A GRanges.
}
\description{
Initialize and run the clustering.
}
