% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{runDeseq2}
\alias{runDeseq2}
\title{Find differentially expressed features.}
\usage{
runDeseq2(object, counts)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{counts}{A \code{list} of \code{vector} of \code{integer}s.}
}
\value{
A \code{list} of feature names.
}
\description{
Find differentially expressed features.
}
