% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{runSlice}
\alias{runSlice}
\title{Initialize and run the slice method.}
\usage{
runSlice(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A GRanges.
}
\description{
Initialize and run the slice method.
}
