% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sampleInfo}
\alias{sampleInfo}
\alias{sampleInfo,srnadiffExp-method}
\title{Accessors for the 'sampleInfo' slot of an srnadiffExp object}
\usage{
\S4method{sampleInfo}{srnadiffExp}(object)
}
\arguments{
\item{object}{An \code{srnadiffExp} object.}
}
\value{
A table containing information on the samples
}
\description{
The \code{sampleInfo} slot holds the sample information as a
\code{data.frame} with three columns labelled \code{FileName},
\code{SampleName} and \code{Condition}. The first column is the
BAM file name (without extension), the second column the sample
name, and the third column the condition to which sample belongs.
Each row describes one sample.
}
\examples{
srnaExp <- srnadiffExample()
sampleInfo(srnaExp)

}
