% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setEmissionProbabilities}
\alias{setEmissionProbabilities}
\alias{setEmissionProbabilities,sRNADiff,numeric-method}
\title{Set emission probabilities (for the HMM step)}
\usage{
setEmissionProbabilities(object, probability)

\S4method{setEmissionProbabilities}{sRNADiff,numeric}(object, probability)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{probability}{The emission probability}
}
\value{
The same object
}
\description{
The emission probabilities follow a binomial distribution for each state.
The second parameter is the probability emitted when the observed p-value
is not less than a given threshold (\code{emissionThreshold}) in the
"not-differentially expressed" state.
This parameter is also the prability emitted when the observed p-value is
not greater than this threshold in the "differentially expressed" state.
}
\examples{
exp <- sRNADiffExample()
exp <- setEmissionProbabilities(exp, 0.9)

}
