% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setEmissionThreshold}
\alias{setEmissionThreshold}
\alias{setEmissionThreshold,sRNADiff,numeric-method}
\title{Set emission threshold (for the HMM step): the emission distribution being
   binomial, all the p-values less than this threshold belong to one class,
   and all the p-values greater than this threshold belong to the
   other class.}
\usage{
setEmissionThreshold(object, threshold)

\S4method{setEmissionThreshold}{sRNADiff,numeric}(object, threshold)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{threshold}{The emission threshold}
}
\value{
The same object
}
\description{
Set emission threshold (for the HMM step): the emission distribution being
   binomial, all the p-values less than this threshold belong to one class,
   and all the p-values greater than this threshold belong to the
   other class.
}
\examples{
exp <- sRNADiffExample()
exp <- setEmissionThreshold(exp, 0.1)

}
