% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{setMergeDistance}
\alias{setMergeDistance}
\title{Set the threshold to merge close regions (in the IR step)}
\usage{
setMergeDistance(object, distance)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{distance}{The maximum distance}
}
\value{
The same object
}
\description{
Set the threshold to merge close regions (in the IR step)
}
\examples{
exp <- sRNADiffExample()
exp <- setMergeDistance(exp, 1000)

}
