% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setMinDepth}
\alias{setMinDepth}
\alias{setMinDepth,sRNADiff,numeric-method}
\title{Set min minimum depth to localize regions}
\usage{
setMinDepth(object, depth)

\S4method{setMinDepth}{sRNADiff,numeric}(object, depth)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{depth}{The minimum depth}
}
\value{
The same object
}
\description{
Control the acceptable minimum coverage to detect a sRNA with
\code{setMinDepth}.
}
\examples{
exp <- sRNADiffExample()
exp <- setMinDepth(exp, 3)

}
