% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setMinDifferences}
\alias{setMinDifferences}
\alias{setMinDifferences,sRNADiff,numeric-method}
\title{Set the threshold to remove similar regions (in the slice step)}
\usage{
setMinDifferences(object, differences)

\S4method{setMinDifferences}{sRNADiff,numeric}(object, differences)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{differences}{The minimum number of different nt.}
}
\value{
The same object
}
\description{
Set the threshold to remove similar regions (in the slice step)
}
\examples{
exp <- sRNADiffExample()
exp <- setMinDifferences(exp, 10)

}
