% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setMinOverlap}
\alias{setMinOverlap}
\alias{setMinOverlap,sRNADiff,numeric-method}
\title{Set minimum overlap (for the last quantification step)}
\usage{
setMinOverlap(object, minOverlap)

\S4method{setMinOverlap}{sRNADiff,numeric}(object, minOverlap)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{minOverlap}{The minimum overlap}
}
\value{
The same object
}
\description{
In the last step, the expression of the potential regions is assessed by
counting the number of reads overlapping each region.
The minimum number of common nucleotides between a read and a region is
controlled by this parameter.
}
\examples{
exp <- sRNADiffExample()
exp <- setMinOverlap(exp, 10)

}
