% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setSizes}
\alias{setSizes}
\alias{setSizes,sRNADiff,numeric,numeric-method}
\title{Set min and max sizes of the regions}
\usage{
setSizes(object, minSize, maxSize)

\S4method{setSizes}{sRNADiff,numeric,numeric}(object, minSize, maxSize)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{minSize}{The minimum size.}

\item{maxSize}{The maximum size.}
}
\value{
The same object
}
\description{
Set min and max sizes of the regions
}
\examples{
exp <- sRNADiffExample()
exp <- setSizes(exp, 10, 1000)
regions(exp)

}
