% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setStrategies}
\alias{setStrategies}
\alias{setStrategies,sRNADiff,logical,logical,logical,logical-method}
\title{Set the different steps}
\usage{
setStrategies(object, annotation, naive, hmm, slice)

\S4method{setStrategies}{sRNADiff,logical,logical,logical,logical}(object,
  annotation, naive, hmm, slice)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{annotation}{The annotation step.}

\item{naive}{The naive step.}

\item{hmm}{The HMM step.}

\item{slice}{The slice step.}
}
\value{
The same object
}
\description{
Set the different steps
}
\examples{
exp <- sRNADiffExample()
exp <- setStrategies(exp, TRUE, FALSE, TRUE, TRUE)

}
