% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{show,sRNADiff-method}
\alias{show,sRNADiff-method}
\title{Overloading the show method}
\usage{
\S4method{show}{sRNADiff}(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A description of the object.
}
\description{
\code{show} prints the parameters used for this experiments, and some
stats of the regions found.
}
\examples{
exp <- sRNADiffExample()
exp

}
