% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{sRNADiff-class}
\alias{sRNADiff-class}
\title{An S4 class to represent sRNA-Seq data for differential expression.}
\description{
An S4 class to represent sRNA-Seq data for differential expression.
}
\section{Slots}{

\describe{
\item{\code{annotation}}{The annotation in GRanges format.}

\item{\code{bamFileNames}}{The name of one read file in BAM format.}

\item{\code{bamFiles}}{The BAM files in a \code{BamFileList}.}

\item{\code{chromosomes}}{The names of the chromosomes.}

\item{\code{chromosomeSizes}}{The sizes of the chromosomes.}

\item{\code{replicates}}{The names of the replicates.}

\item{\code{conditions}}{The condition to which each replicate belongs.}

\item{\code{coverages}}{The coverages, a vector of \code{RLE}.}

\item{\code{lengths}}{The lengths parts of the coverages.}

\item{\code{values}}{The values parts of the coverages.}

\item{\code{design}}{Experimental design, a \code{DataFrame} for
\code{DESeq2}}

\item{\code{pValue}}{The adjusted p-value threshold}

\item{\code{skipAnnotation}}{Whether to skip the annotation strategy step}

\item{\code{skipNaive}}{Whether to skip the naive strategy step}

\item{\code{skipHmm}}{Whether to skip the HMM strategy step}
}}

