% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srnadiff.R
\name{srnadiff}
\alias{srnadiff}
\title{Find differentially expressed sRNA regions}
\usage{
srnadiff(object, segMethod = "all", useParameters, nThreads = 1)
}
\arguments{
\item{object}{An \code{\link{srnadiffExp}} object.}

\item{segMethod}{A character vector. The segmentation methods to use,
one of \code{'annotation'}, \code{'naive'},
\code{'hmm'}, \code{'slice'} or combinations thereof.
Default \code{'all'}, all methods are used. See
Details.}

\item{useParameters}{A named list containing the methods parameters to use.
If missing, default parameter values are supplied.
See \code{\link{parameters}} for details.}

\item{nThreads}{\code{integer(1)}. Number of workers.
Defaults to all cores available as determined by
\code{\link[BiocParallel]{multicoreWorkers}}.}
}
\value{
An \code{srnadiffExp} object containing additional slots for:
\itemize{
    \item \code{regions}
    \item \code{parameters}
    \item \code{countMatrix}
}
}
\description{
This is the main wrapper for running several key functions from this
package.  It is meant to be used after that a \code{\link{srnadiffExp}}
object has been created. \code{\link{srnadiff}} implement four methods to
produce potential DERs (see Details).
Once DERs are detected, the second step in \code{\link{srnadiff}} is to
quantify the statistic signification of these.
}
\details{
Implemented methods to produce potential differentially expressed
regions in \code{\link{srnadiff}} are:

\describe{
\item{\code{annotation:}}{This method simply provides the genomic regions
    corresponding to the annotation file that is optionally given by the
    user. It can be a set of known miRNAs, siRNAs, piRNAs, genes, or a
    combination thereof.}

\item{\code{hmm:}}{This approach assumes that continuous regions of RNA
      along the chromosome are either "differentially expressed" or "not".
      This is captured with a hidden Markov model (HMM) with binary latent
      state of each nucleotide: \emph{differentially expressed} or
      \emph{not differentially expressed}. The observations of the HMM are
      then the empirical p-values arising from the differential expression
      analysis corresponding to each nucleotide position.
      The HMM approach normally needs emission, transition, and starting
      probabilities values (see \code{\link{parameters}}). They
      can be tuned by the user. In order to finding the most likely sequence
      of states from the HMM, the Viterbi algorithm is performed. This
      essentially segments the genome into regions, where a region is
      defined as a set of consecutive bases showing a common expression
      signature.}

\item{\code{slice:}}{In this approach, for each base, the average from
      the normalized coverage is calculated across all samples into each
      condition. This generates a vector of (normalized) mean coverage
      expression per condition. These two vectors
      are then used to compute per-nucleotide log-ratios (in absolute value)
      across the genome. For the computed log-ratio expression, the
      method uses a sliding threshold \emph{h} that run across the log-ratio
      levels identifying bases with log-ratio value above of \emph{h}.
      Regions of contiguous bases passing this threshold are then analyzed
      using an adaptation of Aumann and Lindell algorithm for irreducibility
      property (Aumann and Lindell (2003)).}

\item{\code{naive:}}{This method is the simplest, gived a fixed threshold
      \emph{h}, contiguous bases with log-ratio expression
      (in absolute value) passing this threshold are then considered as
      candidate differentially expressed regions.}
}
}
\examples{
srnaExp <- srnadiffExample()
srnaExp <- srnadiff(srnaExp)
srnaExp

}
\references{
Aumann Y. and, Lindell Y. (2003). A Statistical Theory for Quantitative
Association Rules. \emph{Journal of Intelligent Information Systems},
20(3):255-283.
}
\seealso{
\code{\link{regions}}, \code{\link{parameters}}, \code{\link{countMatrix}}
and \code{\link{srnadiffExp}}
}
