% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{srnadiffDefaultParameters}
\alias{srnadiffDefaultParameters}
\title{List of srnadiff default parameters.}
\format{An object of class \code{list} of length 12.}
\usage{
srnadiffDefaultParameters
}
\description{
List of srnadiff default parameters.
}
\examples{
srnaExp <- srnadiffExample()
srnaExp <- srnadiff(srnaExp, useParameters=srnadiffDefaultParameters)

}
\keyword{data}
