% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{sRNADiffExp}
\alias{sRNADiffExp}
\title{Constructor.}
\usage{
sRNADiffExp(annotation = NULL, bamFileNames, replicates, conditions,
  coverages = NULL, lazyload = FALSE)
}
\arguments{
\item{annotation}{The GRanges annotation}

\item{bamFileNames}{The name of one read file in BAM format.}

\item{replicates}{The names of the replicates.}

\item{conditions}{The condition to which each replicate belongs.}

\item{coverages}{The coverages, a vector of \code{RLE}.}

\item{lazyload}{Usual for S4 functions.}
}
\value{
An \code{sRNADiff} object.
}
\description{
Constructor.
}
\examples{
dir         <- system.file("extdata", package="srnadiff", mustWork = TRUE)
data        <- read.csv(file.path(dir, "data.csv"))
gtfFile     <- file.path(dir, "Homo_sapiens.GRCh38.76.gtf.gz")
annotation  <- readWholeGenomeAnnotation(gtfFile)
bamFiles    <- file.path(dir, data$FileName)
replicates  <- data$SampleName
conditions  <- factor(data$Condition)
exp         <- sRNADiffExp(annotation, bamFiles, replicates, conditions)

}
