% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{srnadiffProcessedExample}
\alias{srnadiffProcessedExample}
\title{Processed example constructor}
\usage{
srnadiffProcessedExample()
}
\value{
An \code{srnadiffExp} object called '\code{srnaExp}'.
}
\description{
This function provides an example of a \code{srnadiffExp} object.
Contrary to \code{srnadiffExample}, this function also runs \code{srnadiff}
on the data with default parameters.
}
\examples{
## The 'srnadiffExp' object in this example was constructed by:

\dontrun{

srnaExp <- srnadiffExample()
srnaExp <- srnadiff(srnaExp)
save(srnaExp, file = file.path(basedir, "srnadiffProcessedExample.rda"))
}

srnaExp <- srnadiffExample()
srnaExp

}
\seealso{
[srnadiff::srnadiffExample()] on which this constructor is based.
}
