% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{buildDataHmm}
\alias{buildDataHmm}
\title{Read the coverage and extract expressed regions}
\usage{
buildDataHmm(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
The selected values: a \code{list} of \code{vector}s of
                  \code{integer}s.
}
\description{
Read the coverage and extract expressed regions
}
