% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeNormalizationFactors}
\alias{computeNormalizationFactors}
\title{Compute normalization factors}
\usage{
computeNormalizationFactors(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
The normalization factor as a list of numeric.
}
\description{
Compute normalization factors
}
