% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRegion}
\alias{plotRegion}
\title{Plot a region}
\usage{
plotRegion(object, region)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{region}{A \code{GenomicRange} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a region
}
\examples{
exp <- sRNADiffExample()
exp <- runAll(exp)
plotRegion(exp, regions(exp, 0.05)[1])

}
