% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_buildHmm}
\alias{rcpp_buildHmm}
\title{Compute unique counts.}
\usage{
rcpp_buildHmm(coverages, minDepth)
}
\arguments{
\item{coverages}{the coverages (a list of RLEs)}

\item{minDepth}{the minimum read coverage}
}
\value{
the unique counts
}
\description{
Compute unique counts.
}
