% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_normalization}
\alias{rcpp_normalization}
\title{Normalize counts}
\usage{
rcpp_normalization(lengths, values, chromosomeSizes, librarySizes)
}
\arguments{
\item{lengths}{the sizes of the RLEs (one list per chromosome)}

\item{values}{the values of the RLEs (one list per chromosome)}

\item{chromosomeSizes}{the sizes of the chromosomes}

\item{librarySizes}{number of elements per sample}
}
\value{
the normalization factors
}
\description{
Normalize counts
}
