% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{readAnnotation}
\alias{readAnnotation}
\title{Segmentation using an annotation file.}
\usage{
readAnnotation(fileName, source = NULL, feature = NULL, name = NULL)
}
\arguments{
\item{fileName}{The annotation file name in GFF/GTF format.}

\item{source}{If not NULL, only lines with this source (2nd field) are
imported.}

\item{feature}{If not NULL, only lines with this feature (3rd field) are
imported.}

\item{name}{If not NULL, use this tag as annotation name.}
}
\value{
A GRanges.
}
\description{
Segmentation using an annotation file.
}
\examples{
dir        <- system.file("extdata", package="srnadiff", mustWork = TRUE)
gtfFile    <- file.path(dir, "Homo_sapiens.GRCh38.76.gtf")
annotation <- readAnnotation(gtfFile, source="miRNA", feature="gene",
                             name="gene_name")

}
