% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{readMiRBasePreAnnotation}
\alias{readMiRBasePreAnnotation}
\title{Segmentation using an miRBase annotation file and use precursor miRNAs.}
\usage{
readMiRBasePreAnnotation(fileName)
}
\arguments{
\item{fileName}{The annotation file name in GFF/GTF format.}
}
\value{
A GRanges.
}
\description{
Segmentation using an miRBase annotation file and use precursor miRNAs.
}
\examples{
dir        <- system.file("extdata", package="srnadiff", mustWork = TRUE)
gffFile    <- file.path(dir, "mirbase21_GRCh38.gff3")
annotation <- readMiRBasePreAnnotation(gffFile)

}
