% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{readWholeGenomeAnnotation}
\alias{readWholeGenomeAnnotation}
\title{Segmentation using an annotation file that contains every genomic feature;
    it extracts the miRNAs.}
\usage{
readWholeGenomeAnnotation(fileName)
}
\arguments{
\item{fileName}{The annotation file name in GFF/GTF format.}
}
\value{
A GRanges.
}
\description{
Segmentation using an annotation file that contains every genomic feature;
    it extracts the miRNAs.
}
\examples{
dir        <- system.file("extdata", package="srnadiff", mustWork = TRUE)
gtfFile    <- file.path(dir, "Homo_sapiens.GRCh38.76.gtf")
annotation <- readWholeGenomeAnnotation(gtfFile)

}
