% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{regions}
\alias{regions}
\alias{regions,sRNADiff,numeric-method}
\alias{regions,sRNADiff,ANY-method}
\title{Get the output regions}
\usage{
regions(object, pvalue = 0.05)

\S4method{regions}{sRNADiff,numeric}(object, pvalue = 0.05)

\S4method{regions}{sRNADiff,ANY}(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{pvalue}{A minimum p-value}
}
\value{
The selected regions
}
\description{
\code{regions} provides the regions previously computed by srnadiff.
}
\examples{
exp <- sRNADiffExample()
regions(exp)

}
