% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runAll}
\alias{runAll}
\title{Run the segmentation using 3 different methods, and reconciliate them.}
\usage{
runAll(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A GRanges.
}
\description{
Run the segmentation using 3 different methods, and reconciliate them.
}
\examples{
dir         <- system.file("extdata", package="srnadiff", mustWork = TRUE)
data        <- read.csv(file.path(dir, "data.csv"))
gtfFile     <- file.path(dir, "Homo_sapiens.GRCh38.76.gtf")
bamFiles    <- file.path(dir, data$FileName)
replicates  <- data$SampleName
conditions  <- factor(data$Condition)
exp         <- sRNADiffExp(gtfFile, bamFiles, replicates, conditions)
diffRegions <- runAll(exp)

}
