% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runAll}
\alias{runAll}
\title{Run the segmentation using 3 different methods, and reconcile them.}
\usage{
runAll(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A \code{GRanges} object.
}
\description{
Run the segmentation using 3 different methods, and reconcile them.
}
\examples{
exp        <- sRNADiffExample()
exp        <- runAll(exp)

}
