% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{runAllAnnotation}
\alias{runAllAnnotation}
\title{Parse an annotation file}
\usage{
runAllAnnotation(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A GRanges.
}
\description{
Parse an annotation file
}
