% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{runAllIR}
\alias{runAllIR}
\title{Segmentation of the genome using a IR method.}
\usage{
runAllIR(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\value{
A \code{GRanges} object.
}
\description{
Segmentation of the genome using a IR method.
}
