% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{runHmm}
\alias{runHmm}
\title{Initialize and run the HMM.}
\usage{
runHmm(object, counts, pvalues)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{counts}{The counts: a \code{list} of \code{vector}s or
\code{integer}s.}

\item{pvalues}{The p-values: a \code{list} of \code{numeric}.}
}
\value{
A GRanges.
}
\description{
Initialize and run the HMM.
}
