% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setEmissionProbabilities}
\alias{setEmissionProbabilities}
\alias{setEmissionProbabilities,sRNADiff,numeric-method}
\title{Set emission probabilities (for the HMM step): probability to have a p-value
   not less than a threshold in the "not-differentially expressed" state,
   and a p-value not greater than this threshold in the "differentially
   expressed" state (supposed equal).}
\usage{
setEmissionProbabilities(object, probability)

\S4method{setEmissionProbabilities}{sRNADiff,numeric}(object, probability)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{probability}{The emission probability}
}
\value{
The same object
}
\description{
Set emission probabilities (for the HMM step): probability to have a p-value
   not less than a threshold in the "not-differentially expressed" state,
   and a p-value not greater than this threshold in the "differentially
   expressed" state (supposed equal).
}
\examples{
exp <- sRNADiffExample()
exp <- setEmissionProbabilities(exp, 0.9)

}
