% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setEmissionThreshold}
\alias{setEmissionThreshold}
\alias{setEmissionThreshold,sRNADiff,numeric-method}
\title{Set emission threshold (for the HMM step)}
\usage{
setEmissionThreshold(object, threshold)

\S4method{setEmissionThreshold}{sRNADiff,numeric}(object, threshold)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{threshold}{The emission threshold}
}
\value{
The same object
}
\description{
The emission probabilities follow a binomial distribution for each state.
Each state, depending whether the p-value is greater or equal to the
second parameter of this function, emits one or another probability.
}
\examples{
exp <- sRNADiffExample()
exp <- setEmissionThreshold(exp, 0.1)

}
