% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setMergeDistance}
\alias{setMergeDistance}
\alias{setMergeDistance,sRNADiff,numeric-method}
\title{Set the threshold to merge close regions (in the naive step)}
\usage{
setMergeDistance(object, distance)

\S4method{setMergeDistance}{sRNADiff,numeric}(object, distance)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{distance}{The maximum distance}
}
\value{
The same object
}
\description{
Set the threshold to merge close regions (in the naive step)
}
\examples{
exp <- sRNADiffExample()
exp <- setMergeDistance(exp, 1000)

}
