% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setMinOverlap}
\alias{setMinOverlap}
\alias{setMinOverlap,sRNADiff,numeric-method}
\title{Set minimum overlap (for the last quantification step): all the reads with
   at least n nucleotides shared with a feature will be used for
   quantification of this feature.}
\usage{
setMinOverlap(object, minOverlap)

\S4method{setMinOverlap}{sRNADiff,numeric}(object, minOverlap)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{minOverlap}{The minimum overlap}
}
\value{
The same object
}
\description{
Set minimum overlap (for the last quantification step): all the reads with
   at least n nucleotides shared with a feature will be used for
   quantification of this feature.
}
\examples{
exp <- sRNADiffExample()
exp <- setMinOverlap(exp, 10)

}
