% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setNThreads}
\alias{setNThreads}
\alias{setNThreads,sRNADiff,numeric-method}
\title{Set number of threads to use}
\usage{
setNThreads(object, nThreads)

\S4method{setNThreads}{sRNADiff,numeric}(object, nThreads)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{nThreads}{The number of threads}
}
\value{
The same object
}
\description{
This only accelerates the \code{DESeq2} calls.
}
\examples{
exp <- sRNADiffExample()
exp <- setNThreads(exp, 4)

}
