% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{setTransitionProbabilities}
\alias{setTransitionProbabilities}
\alias{setTransitionProbabilities,sRNADiff,numeric,numeric-method}
\title{Set transition probabilities (for the HMM step).}
\usage{
setTransitionProbabilities(object, noDiffToDiff, diffToNoDiff)

\S4method{setTransitionProbabilities}{sRNADiff,numeric,numeric}(object,
  noDiffToDiff, diffToNoDiff)
}
\arguments{
\item{object}{An \code{srnadiff} object.}

\item{noDiffToDiff}{probability to change from the "not-differentially
expressed" state to the "differentially expressed"
state}

\item{diffToNoDiff}{probability to change from the "differentially
expressed" state to the "not-differentially expressed"
state}
}
\value{
The same object
}
\description{
Set transition probabilities (for the HMM step).
}
\examples{
exp <- sRNADiffExample()
exp <- setTransitionProbabilities(exp, 0.001, 0.000001)

}
