% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{show,sRNADiff-method}
\alias{show,sRNADiff-method}
\title{Overloading the show method}
\usage{
\S4method{show}{sRNADiff}(object)
}
\arguments{
\item{object}{An \code{srnadiff} object.}
}
\description{
Overloading the show method
}
\examples{
dir         <- system.file("extdata", package="srnadiff", mustWork = TRUE)
data        <- read.csv(file.path(dir, "data.csv"))
gtfFile     <- file.path(dir, "Homo_sapiens.GRCh38.76.gtf")
annotation  <- readWholeGenomeAnnotation(gtfFile)
bamFiles    <- file.path(dir, data$FileName)
replicates  <- data$SampleName
conditions  <- factor(data$Condition)
exp         <- sRNADiffExp(annotation, bamFiles, replicates, conditions)
exp

}
