% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{sRNADiff-class}
\alias{sRNADiff-class}
\title{An S4 class to represent sRNA-Seq data for differential expression.}
\description{
An S4 class to represent sRNA-Seq data for differential expression.
}
\section{Slots}{

\describe{
\item{\code{annotation}}{The annotation in GRanges format.}

\item{\code{bamFileNames}}{The name of one read file in BAM format.}

\item{\code{bamFiles}}{The BAM files in a \code{BamFileList}.}

\item{\code{chromosomes}}{The names of the chromosomes.}

\item{\code{chromosomeSizes}}{The sizes of the chromosomes.}

\item{\code{replicates}}{The names of the replicates.}

\item{\code{conditions}}{The condition to which each replicate belongs.}

\item{\code{coverages}}{The coverages, a vector of \code{RLE}.}

\item{\code{logFC}}{The log fold change.}

\item{\code{design}}{Experimental design, a \code{DataFrame} for
\code{DESeq2}}

\item{\code{regions}}{A \code{GenomicRanges} of the possibly differentially
expressed region}

\item{\code{minDepth}}{Minimum depth to consider to find regions}

\item{\code{minSize}}{Minimum region size}

\item{\code{maxSize}}{Maximum region size}

\item{\code{mergeDistance}}{Distance to merge consecutive region}

\item{\code{noDiffToDiff}}{Transition probability}

\item{\code{diffToNoDiff}}{Transition probability}

\item{\code{emission}}{Emission probability}

\item{\code{emissionThreshold}}{Emission threshold}

\item{\code{skipAnnotation}}{Whether to skip the annotation strategy step}

\item{\code{skipHmm}}{Whether to skip the HMM strategy step}

\item{\code{skipIR}}{Whether to skip the IR strategy step}

\item{\code{nThreads}}{Number of threads}
}}

