% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srnadiff.R
\docType{package}
\name{srnadiff}
\alias{srnadiff}
\alias{srnadiff-package}
\title{srnadiff: A package for differential expression of sRNA-Seq.}
\description{
The srnadiff package provides uses four strategies to find differentially expressed loci.
}
\author{
Matthias Zytnicki, \email{matthias.zytnicki@inra.fr}
}
