% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{sRNADiffExample}
\alias{sRNADiffExample}
\title{Example constructor}
\usage{
sRNADiffExample()
}
\value{
An \code{srnadiff} object
}
\description{
Example constructor
}
\examples{
exp <- sRNADiffExample()

}
