% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{sRNADiffExp}
\alias{sRNADiffExp}
\title{Constructor.}
\usage{
sRNADiffExp(gtfFileName = NA_character_, bamFileNames, replicates, conditions,
  lazyload = FALSE)
}
\arguments{
\item{gtfFileName}{The name of the annotation in GTF format.}

\item{bamFileNames}{The name of one read file in BAM format.}

\item{replicates}{The names of the replicates.}

\item{conditions}{The condition to which each replicate belongs.}

\item{lazyload}{Usual for S4 functions.}
}
\value{
An \code{sRNADiff} object.
}
\description{
Constructor.
}
